Use a 3-arg filter() for a Datastore query
The code for the solution for the Datastore query that is supposed to return
only a specific user's books fails to return any books. The problem is that
the comparison operator and property key were included in a single string.
Break the values out into separate strings so that the filter() method has
a property, comparison operator, and value as expected.

Closes #69.

Signed-off-by: Dan Scott <denials@gmail.com>